am261x: cortex-r-mpu: add MPU region support
The MPU region support is reused from arm_mpu_v7m. a set of new macros
are defined utilizing the existing ones for readability in the file.

CONFIG_ARM_MPU=y in the board.dts will enable the CMakeLists to include
mpu_regions.c defined in the soc/ti/am26x/am261x folder. For changes to
the MPU configurations, the User needs to update this or create new file
and include in the build.

Add Core-build-based differences for the OCRAM MPU regions. DualCore /
LockStep mode differences for TCM.

To operate in lockstep mode, user must set the CORES_IN_LOCKSTEP_MODE
Kconfig symbol to y

Signed-off-by: Mihira Madhava Bollapragada <madhava@ti.com>
Signed-off-by: Shreyas Shankar <s-shankar@ti.com>